neo4j_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
neo4j==5.15.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
openai==1.12.0
requests>=2.31.0
pypdf>=4.0.0
